+++ openfl-workspace/torch/mnist_with_keras/requirements.txt
@@ -1,3 +1,5 @@
 keras==3.8.0
-torch==2.4.1
-torchvision==0.19.1
+torch==2.6.0
+torch-xla==2.6.0
+torchvision==0.21.0
+tensorflow==2.18.0
