+++ openfl/federated/data/__init__.py
@@ -5,13 +5,6 @@
 """Data package."""
 
 from importlib import util
-from warnings import catch_warnings, simplefilter
-
-with catch_warnings():
-    simplefilter(action="ignore", category=FutureWarning)
-    if util.find_spec("tensorflow") is not None:
-        # ignore deprecation warnings in command-line interface
-        import tensorflow  # NOQA
 
 from openfl.federated.data.loader import DataLoader  # NOQA
